Fix size of MDDropdownMenu
The old code assumes that the size of each MDMenuItem or MDMenuItemIcon is dp(56). This is correct for MDMenuItemIcon, but MDMenuItem has a size of dp(48). This caused the menu to be bigger than expected when using MDMenuItem.

The new code calculates the menu's target height taking into consideration the type of item that is being used. It would be ideal if the code could check the height of each item, but it is possible for set_menu_properties to be called before create_menu_items, meaning that the items would not have been created yet.