Toggle button bugfix, Test app addition
The togglebutton module had a bug in wich the buttons created with this
behavior didn't changed even if they where in a group, the colors got
stuck and the overall selection of them was flawn.

This fix uses a fbind to the `state` property, so everytime is changed,
the funciton will be called without being overwritten by the on_release
of the new modules. and also updates the selection if it's state is
changed from outside the class.

`__is_filled` is a boolean memory that turns to true if the custom
object has background filling, this allows to change the font color of
the custom widgets when they're pressed down, the color chagnes to white
and turnd back to the theme_cls.primary_color once it's state is normal.

for the moment this fixes are strict/rigid, but in future revisions it
might become more flexible around customized themes.

MDToggleButton.theme_cls has been added. to save some flops everytime
the MDApp.get_running_app is called to retrieve the app.theme_cls.

At the end of the module was added an ifmain codeblock to allow future
developers to test out the changes faster and in a safe way.

This TEST app uses the instance of the kivymd installed in the
developers pc but the module tested (in this case MDtogglebutton) is the
one in the file.

the module has been documented to help future developres know what's
happening in every line that's not intituive enough.

Some inline comments has been added past the char 80 to allow a better
understanding of the code documented.