Update 11-09-2020 DMY
Changes to be committed: modified: kivymd/uix/button.py

The button's label is now a post processing object, this will allow to
keep a simplified instance for every button, it will create a new instance only when it's needed.

MDicon has been improved!!!
Now the MDicon has the canvas instructiongroup that was inside the class icon_behavior, this is made to keep the button.py code as
clean and explisit as possible.

Instructions and propertyes where removed from icon_behavior and moved to MDIcon

<baseroundbutton> kvlang instructions is removed, no longer needed.

icon comparations are now built inside the MDicon instead of the button.

TODO the code stills needs cleanup.
it will be done later, as the mdiconbutton is fixed, because the mdicon is not showing property inside the
button, theres no problem when the button uses an icon, it only happens when the mdicon uses a texture.