Elevaiton elevated to a new level!
Good news everyone, the elevation bug in wich the widget._elevation
was different than widget.elevation is now solved.

This caused different problems like on elevation 0 the widget shown
a shadow and after that to any other value there was not shadow drawn
and so on, since it was kept the last value before, this means that if we had
elevation = 10
the current elevation of the widget would remain the last elevation in memory
and only after elevation changed, elevation would become the previous elevation
_elevation = 0

example
@starts
_elevation=0
elevation=5
_elevation >> 5

@ update elevation 1
_elevation == 5
elevation = 10
_elevation >> 5

@ update elevation 2
_elevation == 5
elevation = 15
_elevation >> 10

this meant that the elevation was being taken after the _elevation property was used
to update the shadow.

This was solved by changing the widget.elevation deffault value to None
and calling shadow_preset, this function is meant to be redefined and called
upon a super method.

This way, anyone can set a preset for the elevation in case theres no
python code nor kivy lang instance of the widget.

By deffault it will set elevaiton to 10 unless the elevation is set by a kv lang rule,
or pythoncode kwargs in __init__.

Both Circular and rectangular elevaiton behavior used the same KV rule, with the
difference of the texture source, this is easely fixed by updating from a generated
empty texture

also, both of them set the same colors when the _elevation was 0, so the code was now
strip from the subclases and added to the common elevation behavior.

thus makung the startup of the common elevation behavior easyert to load in memory.

if the widget is not an instance of the available sub classes, the update_shadow call will
raise a NotImplementedError.

and simplifying any kv lang overwritting.

also, the kvrule is has now a filename, this means that can now be profiled by the kivy's
kvlang profilling tool.