Selection Control updated!!!
The new MDCheckbox uses the lastest updates to the Buttons.py definitions, as well the
MDicon built in animation!

This update replaces the old MDCheckbox with a new definition that relly in the properties and not
in the kvlang structure, thus allowing a better performance between kvlang definition and python code
instances.

The new behavior and bound box of the button is set inside the MDIconbutton
porting all inside properties to the MDcheckbox.

animate property is added, this property will allow or skip the icon changing animation

Old names are now deprecated and replaced with consistent ones.

the deprecated properties will still work until next big update, until then they will show a
deprecated message in temrinal.