Great Update, MDToggle button now works with the improved button.py classes
this update includes a revition of the kitchensink demo
Black, isort and Flake8 wehre where executed with all files inside the Demo App.

The Toggle button now is a custom class that has basebutton base to allow the widget to overwrite the after_init function.
this helps to avoid any duplicated configuration.

for all configurations if the widget has no normal_version, the __after_init__ function will assign the current value to the normal state.