New update!!!
Fixed position of the label in icon-label buttons (for some reazon pos_hint was being skipt)

bug fix in kitchen sink - the bottom app toolbar was being hack instead of using the proper property (notch = True|False), this meant that the notch space was being drawn when it wasn't needed

Black | Isort | Flake8 checked all files in the repo

NEW!!!

icon buttons now have the property icon_size, this property will control the size of the icon and it will no longer be binded to the normal font_size property.
this means that all developers will be able to change it's size at will.

hurray!!!

the elevatoin of every button is set to 10 if the developer didnt settled the elevation property while creating a new instance.

 Changes to be committed:
	modified:   demos/kitchen_sink/libs/baseclass/bottom_app_bar.py
	modified:   demos/kitchen_sink/libs/baseclass/bottom_sheet.py
	modified:   demos/kitchen_sink/libs/baseclass/dialogs.py
	modified:   demos/kitchen_sink/libs/baseclass/navigation_drawer.py
	modified:   demos/kitchen_sink/libs/baseclass/refresh_layout.py
	modified:   demos/kitchen_sink/main.py
	modified:   kivymd/theming.py
	modified:   kivymd/uix/button.py
	modified:   kivymd/uix/toolbar.py