Elevation behavior (#822)
* Elevation update

Manuel Ortiz

- #794

Shadow generation in real time, it will depend on of the widget size and
shape.

RoundedRectangularElevationBehavior class added, this class is in
charge to allow developer to draw rounded rectangles efficiently for
both shadows.

CommonElevationBehavior is now in charge of every math process that is
shared between different shadows.

1. "RectangularElevationBehavior"
2. "CircularElevationBehavior"
3. "RoundedRectangularElevationBehavior"

Larger widgets will cast a bigger shadow, thus a bigger texture.

Added Rotation support:
 1. Hard shadow shares origin with the widget
 2. Soft shadow has his own origin. Improving the mock up of a real
   shadow work better.

Fake light sourcess support added

Shadow Groups added

Canvas sort support added. (canvas instruction can be found by group
inside `canvas.before`)

To set a new shadow draw function, set the function inside _draw_shadow.
This allows developers to use external functions to draw the shadow
over the canvas.

BackgroundColorBehavior now has rotation support. this allows
independent rotation in order to use:
```py
canvas:
    PushMatrix:
    Trans:
    instr:
    instr:
    instr:
    PopMatrix:
```
Instead of stacking the whole canvas inside the matrix.

DO NOT STACK EVERYTHING FORM DIFFERENT LAYERS.
For example:
```py
canvas.before:
    PushMatrix:
    Trans:
    instr:
    instr:
    instr:
canvas:
    PopMatrix:
```
Is now forbidden because this means that it will take an existing
transformation of the canvas and over do it,  for example, if it's
supposed to have a determined rotation, it will rotate with the new
transformation and then again with the original transformation.
Thus generating visual errors.

Angle is now a reserved word of the background and elevation behaviors
class.
Both classes share the same property and acts independently.

Radius is now a VariableListProperty, this allows an easy set to
every developer.

Cambios a ser confirmados:
	Modificado:     kivymd/uix/behaviors/backgroundcolor_behavior.py
	modificado:     kivymd/uix/behaviors/elevation.py

RoundedRectangularElevationBehavior added to uix/behaviors/__init__.py

Fixed Missing bind to _elevation.

Update documentation of elevation behavior

* Elevation Kitchen sink Demo

The new elevation class and shaders is now documented and added to the 
kitchen sink demo app

* Included improved old shadow definitions.

The old clases for shadow casting are back, but now are called 
FakeRectangularElevationBehavior and FakeCircularElevationBehavior 

Size and transparency are now fixed
The index of the elevation can accept a higher index than 23 (the last 
shadow in the atlas for shadows in both classes)

* bugfix correction

Navigation drawer was taking control from the elevation behavior class, 
causing some bugs in screen and calling re-draws when is unnecesary

* Bug Fix cards now draw with the correct shadow class

* Bugfix and corrections

The BaseElevationButton class was not respecting the Material IO 
standard.
It states in it's examples that every elevated button shoud have 6 
levels of elevation between it's resting state and its raised state.

Bugfix:
The button class was taking control form the elevation behavior class 
thus re drawing the button multiple times.

* Added new controls in the demo screen

This controls are meant to allow users to play with the widgets visual 
properties.
Radius and elevation.

* Improved Shadows and layout for the demo

* Updated docs and values

Included a section on how to properly animate the elevation of the 
widget with the elevaiton behavior.

some bug fixes and optimizations.

* Update of the elevatoin demos

Included the fake shadows to the demo screen

Support for a future upgrade included, not implemented yet.

* Removed development flag

* Improved performance.

* Pre-commit run commit

* Merge branch with master update

Both files have been merged to add the spell check that HeaTTheatR did.

Thanks for your hard work bro, and sorry.

* Merged Demo with master

* Pre-commit run