Button elevation bugfix
Fixed some minor bugs

Fixed animation bugs

Fixed deffault elevation for elevated buttons (min 1dp)
elevation_raised now deppends on widget.elevation

If a elevated button is set to 0, the `__init__` call will set it to
2dp.

All raised elevations will be `widget.elevation` + 6

fixed bug where if the button was animated when it's normal
elevation changed the shadow tilted.

Fixed elevated buttons error where the elevation was broken if a button
was set to disabled from the beginning and then set to eneabled

removed elevation_normal. it's replaced by self.elevation.
because it's behavior was a duplicated value of self.elevation.