WIP - MD Tab Update (#887)
* removed unused kv code

* added some iteration between scroll behavior and fixed behavior

this applies to the tab scroll view. if the tabs width is minor or equal 
to the scroll view, the scroll will be fixed.
If not, will be free to scroll and add a padding of 52 dp horizontally 
as the spec requires.

* Fixed small details of the label configurations.

* removed forced boldness of the label

* Fix for allow_stretch bug

added `fixed_tab_label_width`
This will set the deffault size of the tab when `allow_stretch` is 
`False`.

* changed unused code because of the font_style

* removed unused code

* fixed max_width issue, set to 150 instead 360

* added unbinding methods to remove widgets

* added font_style capability to the MDTab Base

* Fix for tabs unable to set title on kvlang

* Fixed wrongly width scrollable area.

this will fix the issue in wich if the tab was in scroll mode and a 
children is removed, the scrollview never updated to paddin 0 if the 
total width of the children was less than the scroll width, because the 
container never removed the padding.

* Fixed issue in wich if a tab was removed, the canvas reset to index 0

This improve the efficiency of the MDTab remove_widget. giving it 
flexibility to accept removal of either MDTabsLabel or a MDTabsBase 
instance.

Should we consider clear_widget?

* Fixed issue in wich removing a tab would set the carousle index to 0

* Black run

* Fixed issue in wich the tabs list was references backwards.

this is caused because the children list works inverted to a normal 
list, the lastest element added to the list is the first on the children 
list.

* Fixed Tabs documentation

* added MDCarousel documentation

* deprecated value added to tabsbase text, use tab_label_text instead

* Fixed issues with the new `tab_label_text` property compatibility.

* Docs update

* bug fix in for referenced text

The issue was that when a ref text was clicked on, the tab label was set 
to down and if the call back removed the widget, the tab_bar had not 
element with state="down"

* Fixed Docs to simplify work and examples.

* Docs Update

* updated switch_tab to allow new features.

by deffect the fucntion will search with the tab_label_text if the 
argument (name_tab) is a String, if it's a MDTabsBase. the tab will be 
chagned to it instead of iterating between all tabs.

* Pre-commit run

* avoids calling a second update event

* Doc update
fixed missing tabs, if the value doesnt exist, it will throw a 
ValueError exception.

* Docs update.

* Docs Update

* updated ussage for the lastest changes.

* run Pre-commit

* Buxfix and improvement.

recursion fix

removed unnecesary size management, migrated from MDtabsbase.text to 
MDtabsbase.tab_label_text.

* missing deffault search value

* Added configuration property.
Increase flexibility to allow the developer decide if the tab label text 
should be or not Capital

* This update keeps track of every font style change in the theme manager.

* run pre-commit

* cleaned binding calls.

* Added tab_hint_x as a Boolean value.

this will not break the max width of 360dp of each tab.

* Run pre-commit