Remove redundant defaults for MDSpinner
MDSpinner object takes in a ColorProperty as 'color' but gets immediately overridden by the 'primary_color' of the application. The solution proposed allows None value for 'color' and hence override it only if it is a None value. Also, any changes in app's primary color will be reflected only if the original color was set as the primary color. The canvas colors are also changed accordingly.