Changed the way to search for disks for `OSX` and `Linux`, `android` in the `show_disks` function of the `MDFileManager` class, added a preview mode for disks
I decided to do a disk search without using the `subprocess` module, the `mount` command is more universal as for me, the new method will work even if there are more than 2 spaces in the disk name. I also added a `preview` mode for disks, and selected icons similar to `folder.png`.