replaced base name root/self.md_bg_color with *._md_bg_color
this change will allow users to easely change the color of the widgets 
wihtout overwritting the previos one, and thus, make it easier to modify 
the status of the widgets by useing the normal state of the button.

for example, if you get a widget in a group, you can set the current 
color to the accent colour of the theme and when it's not active, just 
update it's color to the base color that is the object.md_bg_color.