COLOR UPDATE :: MDBACKDROP :: fixed color and canvas issues.
* The back layer has been resized to fit correctly the provided space 
behind the front layer. it's also aligned with the top property.

* some documentation has been added to allow developers understand how 
do the MDBackDrop widget works.
* Added documentation for the elevation property, this one will be 
parsed to the front layer, since it requires a Fake elevation behavior 
to display anything.
* The backdrop now works inside a Stencil view, this has been done to 
improve performance, as it will clip anything that is not needed and 
it's outside the render view.
* Improved theming options
* back_layer_color is now deprecated and will be removed soon. this has 
been done in pos of the use of the md_bg_property, that already exists 
and has all the theming logic bound.
* The toolbar background color is now bound to root._md_bg_color
* Backlayer is now a Grid Layout instead of the box layout that was 
previously, this changed because the Grid layout already fits the 
content from top to bottom and from left to right.