MDDatePicker: Change range mode behaviour
Now, after we have selected the date range, we can still change it. Clicking on the already selected end of the range removes the selection from this end, after which we can select another date. Clicking on another date moves the nearest end of the range to this date. Selecting one end of a range means selecting a range that is one day long. This is no longer a forbidden range length. It is no longer forbidden to select the minimum and maximum date in reverse order.