fix hanging the item problem
If only one MDNavigationRailItem was added, it would appear hanging far above the bottom of the Rail.
This is mainly because the internal MDBoxLayout widget that's supposed to contain the MDNaviagtionTrail widget is preset to a fixed sized to allocate space for 3 MDNavigationItems in case they are added, which is not the case some times so it leaves the extra space unoccupied and thus causing a single added Item to appear hanging.

This PR fixes that by allocating space according to the number of MDNavigationItems added so that the space is only enough for those Items and thus eliminating the hanging Items