Prevent triggering events on load
The dispatch calls in the set_current_selected_item method to the triggers on_press and on_release are triggering these events as soon as the widget is added to the navigation rail. To fix this I propose only running the relevant code for highlighting the navigation item contained in those methods. The bug is found on lines 1228 and 1229 of the current version.